\name{formatDF}
\alias{formatDF}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
  Format a Data Frame or Matrix for LaTeX or HTML.
}
\description{
  Format a Data Frame or Matrix for LaTeX or HTML.
}
\usage{
formatDF(...)
## please see ?Hmisc::format.df
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{\dots}{
Please see \code{\link[Hmisc]{format.df}} for details.
}
}
\details{
Alias for the \code{Hmisc::format.df} function whose name will soon be deprecated.
}
\value{
  A character matrix with character images of properly rounded \code{x}.
  Please see \code{\link[Hmisc]{format.df}} for details.
}
\author{
  Frank E. Harrell, Jr.,
  \cr
  Department of Biostatistics,
  \cr
  Vanderbilt University,
  \cr
  \email{f.harrell@vanderbilt.edu}


  Richard M. Heiberger,
  \cr
  Department of Statistics,
  \cr
  Temple University, Philadelphia, PA.
  \cr
  \email{rmh@astro.ocis.temple.edu}


}
\keyword{utilities}
