\name{microplot-internal}
\title{Internal microplot functions}
\alias{optionsCmds}
\alias{sys}
\alias{plot_grid}
\alias{get_legend}
\alias{latex}
\alias{dvi}

\description{Internal microplot functions.}
\details{\code{optionsCmds} and \code{sys} and
  \code{latex} and \code{dvi} are imported from \pkg{Hmisc}. See
  \code{\link[Hmisc]{sys}}.

  \code{plot_grid} and \code{get_legend} are imported from
  \pkg{cowplot}. See \code{\link[cowplot]{plot_grid}} and \code{\link[cowplot]{get_legend}}.
}
\author{ Richard M. Heiberger <rmh@temple.edu> }
\keyword{internal}
