\name{toxicity}
\alias{toxicity}
\docType{data}
\title{
  Simulated toxicity data.  Dataset is used in demo("tablesPlusGraphicColumn").

}
\description{
Simulated toxicity data.  Used in \code{demo("tablesPlusGraphicsColumn", package="microplot")}.
  The demo shows a likert plot \code{\link[HH]{likert}} embedded in a
table of numbers in both 'LaTeX' and 'MS Word'.
}
\usage{data("toxicity")}
\format{
  A data frame with 4 observations on the following 5 variables.
  \describe{
    \item{\code{Grade1}}{a numeric vector}
    \item{\code{Grade2}}{a numeric vector}
    \item{\code{Grade3}}{a numeric vector}
    \item{\code{Grade4}}{a numeric vector}
    \item{\code{Grade5}}{a numeric vector}
  }
}
\author{ Richard M. Heiberger <rmh@temple.edu> }
\examples{
## see demo("tablesPlusGraphicsColumn", package="microplot")
}
\keyword{datasets}
